implemented the flow for the user and products

+++ src/product/product.controller.ts
@@ -1,4 +1,16 @@
-import { Controller, Delete, Get, Param, Post, Put } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  Param,
+  Post,
+  Put,
+  Query,
+} from '@nestjs/common';
+import { isNumber, isString } from 'lodash';
+import { SendResponse } from 'src/utils/common';
+import * as validator from 'validator';
 import { productService } from './product.service';
 
 @Controller('/product')
@@ -6,26 +18,75 @@ export class productController {
   constructor(private readonly productService: productService) {}
 
   @Get()
- async getProduct() {
-    console.log('get product controller reached');
-    return this.productService.getProduct();
+  getProduct(@Query('page') page: string) {
+    return this.productService.getProduct(page);
   }
+
   @Get(':productId')
   getProductById(@Param('productId') productId: string) {
+    if (validator.isEmpty(productId) || !isString(productId))
+      return SendResponse.INVALID_PRODUCT_ID;
+
     return this.productService.getProductById(productId);
   }
+
   @Post()
-  createProduct() {
-    return this.productService.createProduct();
+  createProduct(
+    @Body('name') name: string,
+    @Body('description') description: string,
+    @Body('productNo') productNo: string,
+    @Body('price') price: string,
+  ) {
+    if (validator.isEmpty(name) || !isString(name))
+      return SendResponse.INVALID_PRODUCT_NAME;
+
+    if (validator.isEmpty(description) || !isString(description))
+      return SendResponse.INVALID_PRODUCT_DESCRIPTION;
+
+    if (validator.isEmpty(productNo) || !isString(productNo))
+      return SendResponse.INVALID_PRODUCT_NO;
+
+    if (!isNumber(price) || price < 10) return SendResponse.MIN_PRICE_ERROR;
+
+    return this.productService.createProduct({
+      name,
+      description,
+      productNo,
+      price,
+    });
   }
 
   @Put(':productId')
-  updateProduct(@Param('productId') productId: string) {
-    return this.productService.updateProduct(productId);
+  updateProduct(
+    @Param('productId') productId: string,
+
+    @Body('name') name: string,
+    @Body('description') description: string,
+    @Body('price') price: string,
+  ) {
+    if (validator.isEmpty(productId) || !isString(productId))
+      return SendResponse.INVALID_PRODUCT_ID;
+
+    if (validator.isEmpty(name) || !isString(name))
+      return SendResponse.INVALID_PRODUCT_NAME;
+
+    if (validator.isEmpty(description) || !isString(description))
+      return SendResponse.INVALID_PRODUCT_DESCRIPTION;
+
+    if (!isNumber(price) || price < 10) return SendResponse.MIN_PRICE_ERROR;
+
+    return this.productService.updateProduct(productId, {
+      name,
+      description,
+      price,
+    });
   }
 
   @Delete(':productId')
   deleteProduct(@Param('productId') productId: string) {
+    if (validator.isEmpty(productId) || !isString(productId))
+      return SendResponse.INVALID_PRODUCT_ID;
+
     return this.productService.deleteProduct(productId);
   }
 }
